+++ models/event.js
@@ -1,0 +1,9 @@
+const mongoose = require('mongoose');
+
+const eventSchema = {
+  name: String,
+  date: Date,
+  description: String
+}
+
+module.exports = mongoose.model('events', eventSchema);
