+++ secureCookie.js
@@ -1,5 +1,7 @@
 const crypto = require("crypto");
-module.exports = class secureCookie {
+
+const {APIKey} = require('./Key.json');
+class secureCookie {
     constructor(key) {
         // Digest key as sha-512 hash
         this.key = crypto.createHash("sha512").update(key).digest().slice(0,32);
@@ -36,3 +38,32 @@ module.exports = class secureCookie {
         return decrypted;
     }
 }
+
+const cookieAuthCheck = (req, res, next) => {
+  const cookieCrypt = new secureCookie(APIKey);
+  // Check and see if auth cookie exists
+  if(!req.cookies.auth) 
+      return res.status(401).redirect('/');
+  
+      try {
+      // Decrypt cookie
+      const TSCookie = cookieCrypt.decrypt(req.cookies.auth);
+      const TS = new Date(TSCookie);
+
+      // check if the TSCookie is expired (24 hours)
+      if(TS.getTime() + 1000*60*60 < new Date().getTime()) {
+          // If expired, delete cookie and send login page
+          res.clearCookie('auth');
+          return res.status(401).redirect('/');
+      }
+      // Update cookie timestamp
+      res.cookie('auth',cookieCrypt.encrypt((new Date()).toString()),{httpOnly:true});
+      next();
+  } catch(ex) {
+      // If cookie is invalid, delete cookie and send login page
+      res.clearCookie('auth');
+      return res.status(401).redirect('/');
+  }
+}
+
+module.exports = {secureCookie, cookieAuthCheck};
